Remove OutInstruction's data field
It makes sense for networks which support arbitrary data to do as part of their
address. This reduces the ability to perform DoSs, achieves better performance,
and better uses the type system (as now networks we don't support data on don't
have a data field).

Updates the Ethereum address definition in serai-client accordingly

+++ processor/ethereum/contracts/contracts/Router.sol
@@ -192,7 +192,7 @@ contract Router {
         _transferOut(nextAddress, transactions[i].coin, transactions[i].value);
 
         // Perform the calls with a set gas budget
-        (uint24 gas, bytes memory code) = abi.decode(transactions[i].destination, (uint24, bytes));
+        (uint32 gas, bytes memory code) = abi.decode(transactions[i].destination, (uint32, bytes));
         address(this).call{
           gas: gas
         }(abi.encodeWithSelector(Router.arbitaryCallOut.selector, code));
